trs: split the artifact ABI out of lower.rs (no behavior change)
The plain-Rust half of lower.rs — callback type aliases, InstEnv /
PlanEnv / RuleSpec and friends, CompiledSched/Exec, FnProtos and the
protos wire codec, AOT_LAYOUT_REV, the AOT-mode guard, and the
edge-SSA/fused-plan types — moves to a new always-built trs_codegen::abi
module; lower (feature llvm) glob-re-exports it so every lower::X path
is unchanged.  Pure code motion.

This is the seam for a slim, LLVM-free artifact runtime: loading a
compiled artifact needs these types and codecs but none of inkwell,
and the LLVM statically linked into libtrs_capi.so costs ~5ms of
static-constructor + page-in tax at every process start.
